# perf: switch to json decoder and reduce heap usage 